+++ packages/gateway/src/gateway-language-model-settings.ts
@@ -86,6 +86,7 @@ export type GatewayModelId =
   | 'openai/gpt-5-codex'
   | 'openai/gpt-5-mini'
   | 'openai/gpt-5-nano'
+  | 'openai/gpt-5-pro'
   | 'openai/gpt-oss-120b'
   | 'openai/gpt-oss-20b'
   | 'openai/o1'